requirements.txt
# Flask + CORS (for local or health check)
flask
flask-cors

# Model + performance
transformers>=4.40.0
torch>=2.1.0

safetensors
sentencepiece

# Translation
deep-translator

# RunPod SDK
runpod

# Utilities
requests

huggingface-hub


accelerate>=0.25.0

python-dotenv
